+++ types/setimmediate/index.d.ts
@@ -18,7 +18,7 @@
  * @return The ID of the macrotask, which can be used to abort
  *         the macrotask with `clearImmediate`.
  */
-declare function setImmediate(callback: (...args: any[]) => void, ...args: any[]): number;
+declare function setImmediate<T extends unknown[]>(callback: (...args: T) => void, ...args: T): number;
 
 /**
  * Aborts the specified macrotask before it's run.
