--- types/setimmediate/ts3.1/tsconfig.json
@@ -1,19 +1,0 @@
-{
-    "compilerOptions": {
-        "module": "commonjs",
-        "lib": ["es6"],
-        "noImplicitAny": true,
-        "noImplicitThis": true,
-        "strictFunctionTypes": true,
-        "strictNullChecks": true,
-        "baseUrl": "../../",
-        "typeRoots": ["../../"],
-        "types": [],
-        "noEmit": true,
-        "forceConsistentCasingInFileNames": true
-    },
-    "files": [
-        "index.d.ts",
-        "setimmediate-tests.ts"
-    ]
-}
