--- types/setimmediate/ts3.1/tslint.json
@@ -1,1 +1,0 @@
-{ "extends": "dtslint/dt.json" }
